Update and rename release_new_user_script.py to add_new_user_script_version_to_archives.py
